pyproject.toml
[tool.poetry]
name = "tts-subnet"
version = "0.1.0"
description = "MyShell TTS Subnet"
authors = ["TC <tc@myshell.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
easydict = "^1.13"
jiwer = "^3.0.3"
openai-whisper = "^20231117"
bittensor = "^6.9.3"
melotts = {path = "MeloTTS"}
wandb = "^0.16.4"
safetensors = "^0.4.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
